Cropping Done, with preview
After the cropping, the cropped image would be saved to user table's
:photo_small

The css definition, #preview, in application.css.sass, needs to be
updated because the current work-around, by setting the max-width to be
a very large number, is not clean.

The correct implementation should remove the max-width limit of img tag,
which is previously set by bootstrap.